LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/controller-on-load.https-expected.txt


PASS controller is set for a controlled document
